deriv_hr_platform/requirements.txt
# Core Agent
google-adk==0.3.0
google-generativeai>=0.8.5
python-dotenv>=1.1.0
pandas>=2.0.0
python-dateutil>=2.8.0
litellm>=1.66.0
psutil>=5.9.0
google-genai>=1.0.0

# Server (serves frontend + agent API)
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Slack Bot (Phase 2 — optional, separate process)
slack-bolt>=1.27.0
